requirements.txt
annotated-types==0.7.0
anyio==4.9.0
automium==0.2.6
automium_web==0.1.1
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
config==0.5.1
cssselect2==0.8.0
deprecation==2.1.0
dnspython==2.7.0
email-validator==2.1.0.post1
et_xmlfile==2.0.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.58.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.40.2
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
gotrue==1.3.1
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
legacy==0.1.7
lxml==5.4.0
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.2.1
pillow==10.2.0
plotly==5.17.0
postgrest==0.10.8
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
realtime==1.0.6
reportlab==4.4.1
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
storage3==0.5.4
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
utils==1.0.2
weasyprint==60.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
WTForms==3.1.2
zopfli==0.2.3.post1
